+++ client/src/App.js
@@ -2,7 +2,7 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Routes, useLocation} from 'react-router-dom';
 import { TransitionGroup, CSSTransition } from 'react-transition-group';
-import Home from "./landing/Home";
+import Landing from "./landing/Landing";
 import Login from "./loginpage/Login";
 import Signup from "./sign-up/Signup";
 import Mainpage from "./mainpage/Mainpage";
@@ -20,7 +20,7 @@ function AnimatedRoutes() {
       <CSSTransition key={location.pathname} classNames="fade" timeout={1000}>
         <div className="route-container">
           <Routes>
-            <Route path="/" element={<Home />} />
+            <Route path="/" element={<Landing />} />
             <Route path="/signup" element={<Signup />} />
             <Route path="/login" element={<Login />} />
             <Route path="/dashboard" element={<Mainpage />} />
