+++ client/src/landing/Landing.js
@@ -1,6 +1,6 @@
 // import required dependencies
 import React from "react";
-import "./Home.css";
+import "./Landing.css";
 import divider from "./divider_strip.png";
 import { Button } from "react-bootstrap";
 import Header from "../shared/Header";
